Change order of release automation steps
This commit changes the order of release automation steps and collapses
steps into a single job so that the following steps happen sequentially:

1. Wait for artifacts to be released
2. Create GitHub release and tag release commit
3. Create next development version commit and push
4. Announce release on GChat

Other automated steps happen afterwards. This ensures that the full docs
build of Antora sees a tag for the release and a commit on the branch
that do not point to the same version number in gradle.properties.

Closes gh-53